requirements.txt
Flask-Cors==4.0.1
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
marshmallow==3.21.2
psycopg2==2.9.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.20
gunicorn==22.0.0
python-dotenv==1.0.0


